+++ doc/changelog.rst
@@ -5,11 +5,19 @@
 Changelog
 =========
 
-2.2.1.dev
----------
+2.2.1.dev0
+----------
 
 Minor improvements
 ~~~~~~~~~~~~~~~~~~
+
+New features
+~~~~~~~~~~~~
+- ``iminuit.cost.BinnedNLL`` and  ``iminuit.cost.ExtendedBinnedNLL`` now support
+  weighted binned data
+
+Bug-fixes
+~~~~~~~~~
 - ``FMin.edm_goal`` now remains unchanged if ``Minuit.hesse`` is run after
   ``Minuit.migrad``